chore: Move enum class of Tendency
